project/requirements.txt
torch>=2.1
numpy>=1.24
PyYAML>=6.0
tqdm>=4.66
matplotlib>=3.8
